Move schedule into its own module
